Refactored the BaseUI Schema state management.
- Create only one instance per Schema object.
- Moving all the components in small and manageale separeate files.
- Fixed all the test case issue reported ealier.
- Removed the 'DepListenerContext' context as there was no need for separate context.
