+++ web/pgadmin/static/js/SchemaView/useSchemaState.js
@@ -7,326 +7,21 @@
 //
 //////////////////////////////////////////////////////////////
 
-import React, {
-  useEffect, useMemo, useReducer, useRef, useState,
-} from 'react';
-import diffArray from 'diff-arrays-of-objects';
+import React, { useEffect, useReducer } from 'react';
+
 import _ from 'lodash';
+
+import { parseApiError } from 'sources/api_instance';
 import gettext from 'sources/gettext';
-import { useIsMounted } from '../custom_hooks';
-import { parseApiError } from '../api_instance';
-import BaseUISchema from './base_schema.ui';
-import DepListener  from './DepListener';
-import { isModeSupportedByField, isObjectEqual, isValueEqual } from './utils';
+
+import { DepListener }  from './DepListener';
 import {
-  minMaxValidator, numberValidator, integerValidator, emptyValidator,
-  checkUniqueCol, isEmptyString
-} from '../validators';
+  getSchemaDataDiff,
+  validateSchema,
+} from './schemaUtils';
 
 
 export const SchemaStateContext = React.createContext();
-
-
-function getCollectionDiffInEditMode(
-  field, origVal, sessVal, keepCid, parseChanges
-) {
-  let change = {};
-
-  const id = field.id;
-  const collIdAttr = field.schema.idAttribute;
-  const origColl = _.get(origVal, id) || [];
-  const sessColl = _.get(sessVal, id) || [];
-  /*
-   * Use 'diffArray' package to get the array diff and extract the
-   * info. 'cid' attribute is used to identify the rows uniquely.
-   */
-  const changeDiff = diffArray(
-    origColl, sessColl || [], 'cid', { compareFunction: isObjectEqual }
-  );
-
-  if(changeDiff.added.length > 0) {
-    change['added'] = cleanCid(changeDiff.added, keepCid);
-  }
-
-  if(changeDiff.removed.length > 0) {
-    change['deleted'] = cleanCid(changeDiff.removed.map((row) => {
-      // Deleted records must be from the original data, not the newly added.
-      return _.find(_.get(origVal, field.id), ['cid', row.cid]);
-    }), keepCid);
-  }
-
-  if(changeDiff.updated.length > 0) {
-    /*
-     * There is a change in collection. Parse further deep to figure
-     * out the exact details.
-     */
-    let changed = [];
-
-    for(const changedRow of changeDiff.updated) {
-      const rowIndxSess = _.findIndex(
-        _.get(sessVal, id), (row) => (row.cid === changedRow.cid)
-      );
-      const rowIndxOrig = _.findIndex(
-        _.get(origVal, id), (row) => (row.cid==changedRow.cid)
-      );
-      const finalChangedRow = parseChanges(
-        field.schema, _.get(origVal, [id, rowIndxOrig]),
-        _.get(sessVal, [id, rowIndxSess])
-      );
-
-      if(_.isEmpty(finalChangedRow)) {
-        continue;
-      }
-
-      /*
-       * If the 'id' attribute value is present, then only changed keys
-       * can be passed. Otherwise, passing all the keys is useful.
-       */
-      const idAttrValue = _.get(sessVal, [id, rowIndxSess, collIdAttr]);
-
-      if(_.isUndefined(idAttrValue)) {
-        changed.push({ ...changedRow, ...finalChangedRow });
-      } else {
-        changed.push({ [collIdAttr]: idAttrValue, ...finalChangedRow });
-      }
-    }
-
-    if(changed.length > 0) {
-      change['changed'] = cleanCid(changed, keepCid);
-    }
-  }
-
-  return change;
-}
-
-function getSchemaDataDiff(
-  topSchema, sessData, mode, keepCid, stringify=false, includeSkipChange=true
-) {
-  const isEditMode = mode === 'edit';
-
-  // This will be executed recursively as data can be nested.
-  let parseChanges = (schema, origVal, sessVal) => {
-    let levelChanges = {};
-    parseChanges.depth =
-      _.isUndefined(parseChanges.depth) ? 0 : (parseChanges.depth + 1);
-
-    /* The comparator and setter */
-    const attrChanged = (id, change, force=false) => {
-      if(isValueEqual(_.get(origVal, id), _.get(sessVal, id)) && !force) {
-        return;
-      }
-
-      change = change || _.get(sessVal, id);
-
-      if(stringify && (_.isArray(change) || _.isObject(change))) {
-        change = JSON.stringify(change);
-      }
-
-      /*
-       * Null values are not passed in URL params, pass it as an empty string.
-       * Nested values does not need this.
-       */
-      if(_.isNull(change) && parseChanges.depth === 0) {
-        change = '';
-      }
-
-      levelChanges[id] = change;
-    };
-
-    schema.fields.forEach((field) => {
-      /*
-       * If skipChange is true, then field will not be considered for changed
-       * data. This is helpful when 'Save' or 'Reset' should not be enabled on
-       * this field change alone. No change in other behaviour.
-       */
-      if(field.skipChange && !includeSkipChange) return;
-
-      /*
-       * At this point the schema assignments like top may not have been done,
-       * so - check if mode is supported by this field, or not.
-       */
-      if (!isModeSupportedByField(field, {mode})) return;
-
-      if(
-        typeof(field.type) === 'string' && field.type.startsWith('nested-')
-      ) {
-        /*
-         * Even if its nested, state is on same hierarchical level.
-         * Find the changes and merge.
-         */
-        levelChanges = {
-          ...levelChanges,
-          ...parseChanges(field.schema, origVal, sessVal),
-        };
-      } else if(isEditMode && !_.isEqual(
-        _.get(origVal, field.id), _.get(sessVal, field.id)
-      )) {
-        /*
-         * Check for changes only if in edit mode, otherwise - everything can
-         * go through comparator
-         */
-        if(field.type === 'collection') {
-          const change = getCollectionDiffInEditMode(
-            field, origVal, sessVal, keepCid, parseChanges
-          );
-
-          if(Object.keys(change).length > 0) {
-            attrChanged(field.id, change, true);
-          }
-        } else {
-          attrChanged(field.id);
-        }
-      } else if(!isEditMode) {
-        if(field.type === 'collection') {
-          const origColl = _.get(origVal, field.id) || [];
-          const sessColl = _.get(sessVal, field.id) || [];
-
-          let changeDiff = diffArray(
-            origColl, sessColl, 'cid', {compareFunction: isObjectEqual}
-          );
-
-          // Check the updated changes,when:
-          // 1. These are the fixed rows.
-          // 2. 'canReorder' flag is set to true.
-          if((
-            !_.isUndefined(field.fixedRows) && changeDiff.updated.length > 0
-          ) || (
-            _.isUndefined(field.fixedRows) && (
-              changeDiff.added.length > 0 || changeDiff.removed.length > 0 ||
-              changeDiff.updated.length > 0
-            )
-          ) || (
-            field.canReorder && _.differenceBy(origColl, sessColl, 'cid')
-          )) {
-            attrChanged(
-              field.id, cleanCid(_.get(sessVal, field.id), keepCid), true
-            );
-            return;
-          }
-
-          if(field.canReorder) {
-            changeDiff = diffArray(origColl, sessColl);
-
-            if(changeDiff.updated.length > 0) {
-              attrChanged(
-                field.id, cleanCid(_.get(sessVal, field.id), keepCid), true
-              );
-            }
-          }
-        } else {
-          attrChanged(field.id);
-        }
-      }
-    });
-
-    parseChanges.depth--;
-    return levelChanges;
-  };
-
-  return parseChanges(topSchema, topSchema.origData, sessData);
-}
-
-function validateCollectionSchema(
-  field, sessData, accessPath, setError
-) {
-  const rows = sessData[field.id] || [];
-  const currPath = accessPath.concat(field.id);
-
-  // Validate duplicate rows.
-  const dupInd = checkUniqueCol(rows, field.uniqueCol);
-
-  if(dupInd > 0) {
-    const uniqueColNames = _.filter(
-      field.schema.fields, (uf) => field.uniqueCol.indexOf(uf.id) > -1
-    ).map((uf)=>uf.label).join(', ');
-
-    if (isEmptyString(field.label)) {
-      setError(currPath, gettext('%s must be unique.', uniqueColNames));
-    } else {
-      setError(
-        currPath,
-        gettext('%s in %s must be unique.', uniqueColNames, field.label)
-      );
-    }
-    return true;
-  }
-
-  // Loop through data.
-  for(const [rownum, row] of rows.entries()) {
-    if(validateSchema(
-      field.schema, row, setError, currPath.concat(rownum), field.label
-    )) {
-      return true;
-    }
-  }
-
-  return false;
-}
-
-function validateSchema(
-  schema, sessData, setError, accessPath=[], collLabel=null
-) {
-  sessData = sessData || {};
-
-  for(const field of schema.fields) {
-    // Skip id validation
-    if(schema.idAttribute === field.id) {
-      continue;
-    }
-
-    // If the field is has nested schema, then validate the child schema.
-    if(field.schema && (field.schema instanceof BaseUISchema)) {
-      // A collection is an array.
-      if(field.type === 'collection') {
-        if (validateCollectionSchema(field, sessData, accessPath, setError))
-          return true;
-      }
-      // A nested schema ? Recurse
-      else if(validateSchema(field.schema, sessData, setError, accessPath)) {
-        return true;
-      }
-    } else {
-      // Normal field, default validations.
-      const value = sessData[field.id];
-
-      let message = null;
-      let validators = [];
-
-      if(field.noEmpty) {
-        const label = (
-          collLabel && gettext('%s in %s', field.label, collLabel)
-        ) || field.noEmptyLabel || field.label;
-
-        validators.push[emptyValidator.bind(null, label, value)];
-      }
-
-      if(field.type === 'int') {
-        validators.push[integerValidator.bind(null, field.label, value)];
-        validators.push[
-          minMaxValidator(null, field.label, value, field.min, field.max)
-        ];
-      } else if(field.type === 'numeric') {
-        validators.push[numberValidator.bind(null, field.label, value)];
-        validators.push[
-          minMaxValidator(null, field.label, value, field.min, field.max)
-        ];
-      }
-
-      for (const validator in validators) {
-        message = validator();
-
-        if (message) {
-          setError(accessPath.concat(field.id), message);
-          return true;
-        }
-      }
-    }
-  }
-
-  return schema.validate(
-    sessData, (id, message) => setError(accessPath.concat(id), message)
-  );
-}
 
 export const SCHEMA_STATE_ACTIONS = {
   INIT: 'init',
@@ -353,7 +48,7 @@ const getDepChange = (currPath, newState, oldState, action) => {
   return newState;
 };
 
-const getDeferredDepChange = (currPath, newState, oldState, action)=>{
+const getDeferredDepChange = (currPath, newState, oldState, action) => {
   if(action.deferredDepChange) {
     return action.deferredDepChange(currPath, newState, {
       type: action.type,
@@ -441,22 +136,17 @@ const sessDataReducer = (state, action) => {
 
   case SCHEMA_STATE_ACTIONS.CLEAR_DEFERRED_QUEUE:
     data.__deferred__ = [];
-    break;
+    return data;
 
   case SCHEMA_STATE_ACTIONS.DEFERRED_DEPCHANGE:
     data = getDepChange(action.path, data, state, action);
     break;
-
   }
+
+  data.__changeId = (data.__changeId || 0) + 1;
+
   return data;
 };
-
-// Remove cid key added by prepareData
-const cleanCid = (coll, keepCid=false) => (
-  (!coll || keepCid) ? coll : coll.map(
-    (o) => _.pickBy(o, (v, k) => (k !== 'cid'))
-  )
-);
 
 function prepareData(val, createMode=false) {
   if(_.isPlainObject(val)) {
@@ -489,209 +179,286 @@ const LOADING_STATE = {
   ERROR: 'Error'
 };
 
-export default function useSchemaState({
-  schema, getInitData, immutableData, mode, keepCid, resetKey,
-  setLoadingMessage, onDataChange,
-}) {
-  immutableData = immutableData || {};
+class SchemaState extends DepListener {
 
-  const res = useMemo(() => ({
-    state: null,
-  }), [resetKey, immutableData]);
+  constructor(
+    schema, getInitData, immutableData, mode, keepCid, onDataChange
+  ) {
+    super();
 
-  const [loadinState, setLoadingState] = useState(LOADING_STATE.INIT);
-  const depListener = useRef(new DepListener());
-  const [errors, setError] = useState({});
-  const isNew = schema.isNew(schema.origData);
-  const checkIsMounted = useIsMounted();
+    ////// Helper variables
 
-  const [isDirty, setDirty] = useState(false);
-  const [isReady, setReady] = useState(false);
-  const preReadyQueue = useRef([]);
+    // BaseUISchema instance
+    this.schema = schema;
+    // Current mode of operation ('create', 'edit', 'properties')
+    this.mode = mode;
+    // Keep the 'cid' object during diff calculations.
+    this.keepcid = keepCid;
+    // Initialization callback
+    this.getInitData = getInitData;
+    // Data change callback
+    this.onDataChange = onDataChange;
 
-  const changes = useRef(null);
+    ////// State variables
 
-  // Schema data
-  const [sessData, sessDispatch] = useReducer(sessDataReducer, {});
+    // Is is ready to be consumed?
+    this.isReady = false;
+    // Diff between the current snapshot and initial data.
+    this.changes = null;
+    // Loading message (if any)
+    this.message = null;
+    // Current Loading state
+    this.loadingState = LOADING_STATE.INIT;
 
-  setLoadingMessage = setLoadingMessage || (() => { /* Do nothing */ });
+    ////// Schema instance data
 
-  // Dispatch all the actions recorded before it was ready.
-  useEffect(() => {
-    if(isReady) {
-      if(preReadyQueue.current.length > 0) {
-        for (const payload  of preReadyQueue.current) {
-          sessDispatch(payload);
-        }
-      }
-      // Destroy the queue so that no one uses it.
-      preReadyQueue.current = undefined;
-    }
-  }, [isReady]);
+    // Initial data after the ready state
+    this.initData = {};
+    // Immutable data
+    this.immutableData = immutableData;
+    // Current error
+    this.errors = {};
+    // Pre-ready queue
+    this.preReadyQueue = [];
 
-  // Validate the schema on the change of the data.
-  useEffect(() => {
-    // If schema does not have the data or does not have any 'onDataChange'
-    // callback, there is no need to validate the current data.
-    if(!isReady || !onDataChange) return;
+    this._id = Date.now();
+  }
 
-    if(
-      !validateSchema(schema, sessData, (path, message) => {
-        message && setError({ name: path, message: _.escape(message) });
-      })
-    ) setError({});
+  setError(err) {
+    this.errors = err;
+  }
 
-    // Check if anything changed.
-    let dataDiff = getSchemaDataDiff(
-      schema, sessData, mode, keepCid, false, false
-    );
-    const hasDataChanged = Object.keys(dataDiff).length > 0;
-    setDirty(hasDataChanged);
+  setReady(state) {
+    this.isReady = state;
+  }
 
-    // Inform the callbacks about change in the data.
-    if(mode !== 'edit') {
-      // Merge the changed data with origData in 'create' mode.
-      dataDiff = _.assign({}, schema.origData, dataDiff);
-    } else {
-      dataDiff[schema.idAttribute] = schema.origData[schema.idAttribute];
-    }
-    changes.current = hasDataChanged ? dataDiff : null;
+  setLoadingState(loadingState) {
+    this.loadingState = loadingState;
+  }
 
-    onDataChange?.(hasDataChanged, dataDiff);
-  }, [sessData, isReady]);
+  setLoadingMessage(msg) {
+    this.message = msg;
+  }
 
-  useEffect(() => {
-    if(sessData.__deferred__?.length > 0) {
-      const items = sessData.__deferred__;
+  // Initialise the data, and fetch the data from the backend (if required).
+  // 'force' flag can be used for reloading the data from the backend.
+  initialise(dataDispatch, force) {
+    let state = this;
 
-      sessDispatch({
-        type: SCHEMA_STATE_ACTIONS.CLEAR_DEFERRED_QUEUE,
-      });
+    // Don't attempt to initialize again (if it's already in progress).
+    if (
+      state.loadingState !== LOADING_STATE.INIT ||
+      (force && state.loadingState === LOADING_STATE.LOADING)
+    ) return;
 
-      items.forEach((item) => {
-        item.promise.then((resFunc) => {
-          sessDispatch({
-            type: SCHEMA_STATE_ACTIONS.DEFERRED_DEPCHANGE,
-            path: item.action.path,
-            depChange: item.action.depChange,
-            listener: {
-              ...item.listener,
-              callback: resFunc,
-            },
-          });
-        });
-      });
-    }
-  }, [sessData.__deferred__?.length]);
-
-  useEffect(() => {
-    if (loadinState !== LOADING_STATE.INIT) return;
-    setLoadingState(LOADING_STATE.LOADING);
-    setLoadingMessage(gettext('Loading...'));
-
+    state.setLoadingState(LOADING_STATE.LOADING);
+    state.setLoadingMessage(gettext('Loading...'));
 
     /*
      * Fetch the data using getInitData(..) callback.
      * `getInitData(..)` must be present in 'edit' mode.
      */
-    if(mode === 'edit' && !getInitData) {
+    if(state.mode === 'edit' && !state.getInitData) {
       throw new Error('getInitData must be passed for edit');
     }
 
-    const initDataPromise = getInitData?.() || Promise.resolve({});
+    const initDataPromise = state.getInitData?.() ||
+      Promise.resolve({});
 
     initDataPromise.then((data) => {
       data = data || {};
 
-      if(mode === 'edit') {
+      if(state.mode === 'edit') {
         // Set the origData to incoming data, useful for comparing.
-        schema.origData = prepareData({...(data || {}), ...immutableData});
+        state.initData = prepareData({...data, ...state.immutableData});
       } else {
         // In create mode, merge with defaults.
-        schema.origData = prepareData({
-          ...schema.defaults,
-          ...data,
-          ...immutableData,
+        state.initData = prepareData({
+          ...state.schema.defaults, ...data, ...state.immutableData
         }, true);
       }
-      schema.initialise(schema.origData);
 
-      // Let's wait for the component to mount first.
-      if(!checkIsMounted()) return;
+      state.schema.initialise(state.initData);
 
-      setReady(true);
-      setLoadingMessage('');
-      sessDispatch({
+      dataDispatch({
         type: SCHEMA_STATE_ACTIONS.INIT,
-        payload: schema.origData,
+        payload: state.initData,
       });
-      setLoadingState(LOADING_STATE.LOADED);
+
+      state.setLoadingState(LOADING_STATE.LOADED);
+      state.setLoadingMessage('');
+      state.setReady(true);
     }).catch((err) => {
-      setReady(true);
-      setError({
+      state.setLoadingMessage('');
+      state.setError({
         name: 'apierror',
         response: err,
         message: _.escape(parseApiError(err)),
       });
-      if(!checkIsMounted()) return;
-      setLoadingMessage('');
-      setLoadingState(LOADING_STATE.ERROR);
+      state.setLoadingState(LOADING_STATE.ERROR);
+      state.setReady(true);
     });
-  }, [immutableData]);
+  }
 
-  useEffect(() => {
-    // If reset key changes, reset the form.
-    schema.initialise(schema.origData);
-    sessDispatch({
-      type: SCHEMA_STATE_ACTIONS.INIT,
-      payload: schema.origData,
-    });
-  }, [resetKey]);
+  validate(sessData) {
+    let state = this,
+      schema = state.schema;
+
+    // If schema does not have the data or does not have any 'onDataChange'
+    // callback, there is no need to validate the current data.
+    if(!state.isReady || !state.onDataChange) return;
+
+    if(
+      !validateSchema(schema, sessData, (path, message) => {
+        message && state.setError({ name: path, message: _.escape(message) });
+      })
+    ) state.setError({});
+
+    // Check if anything changed.
+    let dataDiff = getSchemaDataDiff(
+      schema, state.initData, sessData,
+      state.mode, state.keepCid, false, false
+    );
+    const hasDataChanged = Object.keys(dataDiff).length > 0;
+
+    // Inform the callbacks about change in the data.
+    if(state.mode !== 'edit') {
+      // Merge the changed data with origData in 'create' mode.
+      dataDiff = _.assign({}, state.initData, dataDiff);
+
+      // Remove internal '__changeId' attribute.
+      delete dataDiff.__changeId;
+    } else if (hasDataChanged) {
+      const idAttr = schema.idAttribute;
+      dataDiff[idAttr] = state.initData[idAttr];
+    }
+
+    state.changes = hasDataChanged ? dataDiff : null;
+
+    state.onDataChange(hasDataChanged, dataDiff);
+  }
+
+  get isNew() {
+    return this.schema.isNew(this.initData);
+  }
+
+  set isNew(val) {
+    throw new Error('Property \'isNew\' is readonly.', val);
+  }
+
+  get isDirty() {
+    return this.changes !== null;
+  }
+
+  set isDirty(val) {
+    throw new Error('Property \'isDirty\' is readonly.', val);
+  }
+}
+
+export const useSchemaState = ({
+  schema, getInitData, immutableData, mode, keepCid, onDataChange,
+}) => {
+  let schemaState = schema.state;
+
+  if (!schemaState) {
+    schemaState = new SchemaState(
+      schema, getInitData, immutableData, mode, keepCid, onDataChange
+    );
+    schema.state = schemaState;
+  }
+
+  const [sessData, sessDispatch] = useReducer(
+    sessDataReducer, {__changeId: 0}
+  );
 
   const sessDispatchWithListener = (action) => {
     let dispatchPayload = {
       ...action,
-      depChange: (...args) => depListener.current.getDepChange(...args),
-      deferredDepChange:
-      (...args) => depListener.current.getDeferredDepChange(...args),
+      depChange: (...args) => schemaState.getDepChange(...args),
+      deferredDepChange: (...args) => schemaState.getDeferredDepChange(...args),
     };
     /*
      * All the session changes coming before init should be queued up.
      * They will be processed later when form is ready.
      */
-    preReadyQueue.current ?
-      preReadyQueue.current.push(dispatchPayload) :
+    let preReadyQueue = schemaState.preReadyQueue;
+
+    preReadyQueue ?
+      preReadyQueue.push(dispatchPayload) :
       sessDispatch(dispatchPayload);
   };
 
-  // Set the _sessData, can be usefull to some deep controls.
-  schema._sessData = sessData;
-
-  if (res.state !== null) return res.state;
-
-  // Schema state
-  res.state = {
-    dataDispatch: sessDispatchWithListener,
-    initOrigData: (path, value) => {
-      if(path) {
-        let data = prepareData(value);
-        _.set(schema.origData, path, data);
-        sessDispatchWithListener({
-          type: SCHEMA_STATE_ACTIONS.SET_VALUE,
-          path: path,
-          value: data,
-        });
-      }
-    },
-    depListener: depListener,
-    changes: changes,
-    sessData,
-    errors,
-    isDirty,
-    isReady,
-    isNew,
-    setError,
+  schemaState.setUnpreparedData = (path, value) => {
+    if(path) {
+      let data = prepareData(value);
+      _.set(schema.initData, path, data);
+      sessDispatchWithListener({
+        type: SCHEMA_STATE_ACTIONS.SET_VALUE,
+        path: path,
+        value: data,
+      });
+    }
   };
 
-  return res.state;
-}
+  const resetData = () => {
+    const initData = _.cloneDeep(schemaState.initData);
+    initData.__changeId = sessData.__changeId;
+    sessDispatch({
+      type: SCHEMA_STATE_ACTIONS.INIT,
+      payload: initData,
+    });
+  };
+
+  useEffect(() => {
+    schemaState.initialise(sessDispatch);
+  }, [schemaState.loadingState]);
+
+  useEffect(() => {
+    let preReadyQueue = schemaState.preReadyQueue;
+
+    if (!schemaState.isReady || !preReadyQueue) return;
+
+    for (const payload  of preReadyQueue) {
+      sessDispatch(payload);
+    }
+
+    // Destroy the queue so that no one uses it.
+    schemaState.preReadyQueue = null;
+  }, [schemaState.isReady]);
+
+  useEffect(() => {
+    // Validate the schema on the change of the data.
+    schemaState.validate(sessData);
+  }, [schemaState.isReady, sessData.__changeId]);
+
+  useEffect(() => {
+    const items = sessData.__deferred__ || [];
+
+    if (items.length == 0) return;
+
+    sessDispatch({
+      type: SCHEMA_STATE_ACTIONS.CLEAR_DEFERRED_QUEUE,
+    });
+
+    items.forEach((item) => {
+      item.promise.then((resFunc) => {
+        sessDispatch({
+          type: SCHEMA_STATE_ACTIONS.DEFERRED_DEPCHANGE,
+          path: item.action.path,
+          depChange: item.action.depChange,
+          listener: {
+            ...item.listener,
+            callback: resFunc,
+          },
+        });
+      });
+    });
+  }, [schemaState.__deferred__?.length]);
+
+  return {
+    schemaState,
+    dataDispatch: sessDispatchWithListener,
+    sessData,
+    reset: resetData,
+  };
+};
